euler/src/main/java/me/victar/euler/Euler.java
package me.victar.euler;

import java.util.HashMap;
import java.util.Map;

public class Euler {
	private final int power;
	private Map<Integer,Integer> powerMap = new HashMap<Integer,Integer>();
	
	public Euler(int power) {
		this.power=power;
	}

	public int getPower() {
		return power;
	}

	public int getPowNumber(int i) {
		return (int) Math.pow(i, power);
	}

	public int[] getCortege(int number) {
		int powNumber = getPowNumber(number);
		for (int i=0; i<number; i++){
			powerMap.put(i,getPowNumber(i));
		}
		
		Map<Integer,Integer> currentMap = new HashMap<Integer,Integer>();

		for (int i = 0; i < power; i++) {
			for (int j = 0; j < number; j++) {
				if (!currentMap.containsKey(i)) {
					currentMap.put(j, powerMap.get(j));
					getCortegeSum(currentMap);
				}
			}
			;
		}
//		for (int i = )
		return  new int[]{3,4};
	}
	
	

	public int getCortegeMinLength() {
		return 2;
	}

	public int getCortegeSum(Map<Integer, Integer> map) {
		int result = 0;
		for (Map.Entry<Integer, Integer> entry : map.entrySet()) {
		   result +=entry.getValue();
		}
		return result;
	}

}
